requirements.txt
boto3
elasticsearch==7.5.1
pandas
requests
uuid
psycopg2
flask
python-dotenv